Conflicts:
	src/main/java/kosta/pro/rgmall/controller/AdminController.java
	src/main/java/kosta/pro/rgmall/domain/Orders.java
	src/main/java/kosta/pro/rgmall/service/UserServiceImpl.java
	src/main/webapp/WEB-INF/views/myPage/userOrderList.jsp
	src/main/webapp/WEB-INF/views/user/myPageForm.jsp
	src/main/webapp/WEB-INF/views/user/order.jsp

	충돌수정

+++ src/main/java/kosta/pro/rgmall/controller/MainController.java
@@ -7,13 +7,18 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
+import kosta.pro.rgmall.domain.Admin;
 import kosta.pro.rgmall.domain.FAQ;
 import kosta.pro.rgmall.domain.GoodsAnswer;
 import kosta.pro.rgmall.domain.GoodsQuestion;
@@ -88,7 +93,7 @@ public class MainController {
 
 		UserList userList = mainService.findUserEmail(userEmail);
 		session.setAttribute("userList", userList);
-		
+
 		mainService.deleteUserByName("kakaoTest");
 
 		return "main/index";
@@ -209,14 +214,16 @@ public class MainController {
 	 * 상품리스트를 조회를 하는 Controller
 	 */
 	@RequestMapping("/goodsList/{main}/{sub}/{sort}")
-	public ModelAndView goodsList(@PathVariable Long main, @PathVariable Long sub, @PathVariable int sort) {
+	public ModelAndView goodsList(@PathVariable Long main, @PathVariable Long sub, @PathVariable int sort,
+			@RequestParam(defaultValue = "0")int nowPage) {
 		Map<String, Object> goodsListMap = new HashMap<String, Object>();
-		List<RegisterGoods> registerGoodsList = mainService.selectAllGoods(main, sub, sort);
-		System.out.println(sort);
+		Pageable pageable = PageRequest.of(nowPage, 16);
+		Page<RegisterGoods> registerGoodsList = mainService.selectAllGoods(main, sub, sort,pageable);
 		List<MainCategories> mainCategories = mainService.selectCategories();
-
+		goodsListMap.put("sort", sort);
 		goodsListMap.put("registerGoodsList", registerGoodsList);
-		goodsListMap.put("registerGoods", registerGoodsList.get(0));
+		goodsListMap.put("regGoods",registerGoodsList.getContent());
+		goodsListMap.put("registerGoods", registerGoodsList.getContent().get(0));
 		goodsListMap.put("mainCategories", mainCategories);
 		goodsListMap.put("main", main);// main
 		goodsListMap.put("sub", sub);// sub
@@ -231,21 +238,24 @@ public class MainController {
 	 */
 	@RequestMapping("/goodsListOrder/{main}/{sub}/{sort}")
 	@ResponseBody
-	public List<RegisterGoods> goodsListOrder(@PathVariable Long main, @PathVariable Long sub, @PathVariable int sort) {
-		List<RegisterGoods> registerGoodsList = mainService.selectAllGoods(main, sub, sort);
-		return registerGoodsList;
+	public List<RegisterGoods> goodsListOrder(@PathVariable Long main, @PathVariable Long sub, @PathVariable int sort,
+			@RequestParam(defaultValue = "0")int nowPage) {
+		Pageable pageable = PageRequest.of(nowPage, 16);
+		Page<RegisterGoods> registerGoodsList = mainService.selectAllGoods(main, sub, sort,pageable);
+		return registerGoodsList.getContent();
 	}
-	
+
 	/**
 	 * 상품 keyword검색하는 Controller
-	 * */
+	 */
 	@ResponseBody
 	@RequestMapping(value = "/goodsListKeyword", method = RequestMethod.POST)
-	public List<RegisterGoods> goodsListKeyword(String keyword) {
-		List<RegisterGoods> registerGoodsList = mainService.searchGoods(keyword);
-		return registerGoodsList;
+	public List<RegisterGoods> goodsListKeyword(String keyword,@RequestParam(defaultValue = "0")int nowPage) {
+		Pageable pageable = PageRequest.of(nowPage, 16);
+		Page<RegisterGoods> registerGoodsList = mainService.searchGoods(keyword,pageable);
+		return registerGoodsList.getContent();
 	}
-	
+
 	/**
 	 * 상품 상세조회를 하는 Controller
 	 */
@@ -264,42 +274,50 @@ public class MainController {
 
 		goodsQuestionMap.put("registerGoods", registerGoods);
 		goodsQuestionMap.put("goodsQuestionList", goodsQuestionList);
-		
+
 		List<Review> review = mainService.selectReview(regNo);
-		goodsQuestionMap.put("review",review);
-		
+		goodsQuestionMap.put("review", review);
+
 		ModelAndView mv = new ModelAndView("main/goodsDetail", "goodsQuestionMap", goodsQuestionMap);
 		return mv;
 	}
-
 
 	/**
 	 * 상품문의 A에 대한 기능(only 관리자)
 	 */
 	@RequestMapping("/goodsDetail/{qgoodsNo}/goodsAnswer/{state}") // 나중에 세션 인수로 받아야할듯?
-	public String UpdateGoodsAnswer(@PathVariable Long qgoodsNo, @PathVariable String state,
-			HttpServletRequest request) {
+	public String UpdateGoodsAnswer(@PathVariable Long qgoodsNo, @PathVariable String state, HttpServletRequest request,
+			HttpSession session) {
 
-		if (state.equals("update")) {
-			String content = request.getParameter("content");
+		Admin sessionUserList = (Admin) session.getAttribute("userList");
+		String authority = sessionUserList.getAuthority();
+		if (authority.equals("ROLE_ADMIN")) {
 
-			GoodsQuestion goodsQuestion = new GoodsQuestion(qgoodsNo);
-			GoodsAnswer goodsAnswer = new GoodsAnswer(content, goodsQuestion);
+			if (state.equals("update")) {
+				String content = request.getParameter("content");
 
-			mainService.updateGoodsAnswer(goodsAnswer);
+				GoodsQuestion goodsQuestion = new GoodsQuestion(qgoodsNo);
+				GoodsAnswer goodsAnswer = new GoodsAnswer(content, goodsQuestion);
 
-		} else if (state.equals("delete")) {
-			mainService.deleteGoodsAnswer(qgoodsNo);
-		} else if (state.equals("insert")) {
+				mainService.updateGoodsAnswer(goodsAnswer);
 
-			String content = request.getParameter("content");
+			} else if (state.equals("delete")) {
+				mainService.deleteGoodsAnswer(qgoodsNo);
+			} else if (state.equals("insert")) {
 
-			GoodsQuestion goodsQuestion = new GoodsQuestion(qgoodsNo);
-			GoodsAnswer goodsAnswer = new GoodsAnswer(content, goodsQuestion);
+				String content = request.getParameter("content");
 
-			mainService.insertGoodsAnswer(goodsAnswer);
+				GoodsQuestion goodsQuestion = new GoodsQuestion(qgoodsNo);
+				GoodsAnswer goodsAnswer = new GoodsAnswer(content, goodsQuestion);
 
-		}
+				mainService.insertGoodsAnswer(goodsAnswer);
+
+			}
+		} /*
+			 * else if (authority.equals("ROLE_USER")) {
+			 * 
+			 * }
+			 */
 
 		return "redirect:/main/goodsList/0/0/0";
 	}
@@ -308,8 +326,8 @@ public class MainController {
 	 * 상품문의 Q에 대한 기능(유저)
 	 */
 	@RequestMapping("/goodsDetail/{regNo}/goodsQuestion/{state}") // 나중에 세션 인수로 받아야할듯?
-	public String UpdateGoodsQuestion(@PathVariable Long regNo, @PathVariable String state,
-			HttpServletRequest request) {
+	public String UpdateGoodsQuestion(@PathVariable Long regNo, @PathVariable String state, HttpServletRequest request,
+			HttpSession session) {
 
 		if (state.equals("update")) {
 
@@ -325,7 +343,10 @@ public class MainController {
 
 			String content = request.getParameter("content");
 
-			UserList userList = new UserList(93L);
+			UserList sessionUserList = (UserList) session.getAttribute("userList");
+			Long userNo = sessionUserList.getUserNo();
+
+			UserList userList = new UserList(userNo);
 			RegisterGoods registerGoods = new RegisterGoods(regNo);
 
 			GoodsQuestion goodsQuestion = new GoodsQuestion(content, userList, registerGoods);
@@ -336,7 +357,7 @@ public class MainController {
 
 		return "redirect:/main/goodsList/0/0/0";
 	}
-	
+
 	/**
 	 * FAQ에서 검색
 	 */
@@ -347,14 +368,12 @@ public class MainController {
 		return faq;
 	}
 
-	
-	
 	/**
 	 * 우편번호 api test
 	 */
 	@RequestMapping("/testtt")
 	public void APITest() {
-		
+
 	}
-	
+
 }// class
